# swagger-app/api-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API documentation of the different services available in the Ramp-Up\
    \ Journal system"
  version: "0.1.0"
  title: "Ramp-up Journal - API"
basePath: "/api"
produces:
- "application/json"
paths:
  /task-definitions/{id}:
    get:
      tags:
      - "task definition"
      summary: "Retrieve a specific task definition by the given id"
      description: "This endpoint will retrieve a task-definition based on the id\
        \ received by parameter"
      operationId: "task_definitionsIdGET"
      parameters:
      - name: "id"
        in: "path"
        description: "Task definition id"
        required: true
        type: "string"
      responses:
        200:
          description: "task-definition details"
          schema:
            $ref: "#/definitions/TaskDefinition"
        403:
          description: "Forbidden, The request was not allowed because the client\
            \ does not have the required credentials"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Task definition not found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "TaskDefinition"
    delete:
      tags:
      - "task definition"
      summary: "Delete a specific task definition given its id"
      operationId: "task_definitionsIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        description: "Task definition id"
        required: true
        type: "string"
      responses:
        200:
          description: "The task definition was deleted successfully or doesn't exist"
        500:
          description: "Unexpected error, The task definition may not have been deleted"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "TaskDefinition"
  /task-definitions:
    get:
      tags:
      - "task definition"
      summary: "Retrieve a list of tasks"
      description: "This endpoint will retrieve a list of task-definitions"
      operationId: "task_definitionsGET"
      parameters: []
      responses:
        200:
          description: "OK, Obtains the complete list of task definitions stored in\
            \ the system"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TaskDefinition"
        500:
          description: "Unexpected system error, it wasn't able to retrieve the list\
            \ of task definitions."
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "TaskDefinition"
    post:
      tags:
      - "task definition"
      summary: "Create a new task definition"
      description: "This endpoint shall be used to create a new task-definition"
      operationId: "task_definitionsPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Task Definition data"
        required: true
        schema:
          $ref: "#/definitions/TaskDefinition"
      responses:
        201:
          description: "The task definition was created successfully"
          schema:
            $ref: "#/definitions/Id"
          headers:
            Location:
              type: "string"
              description: "The URI or path of the created task definition."
        400:
          description: "Bad Request, The request body does not conform to a valid\
            \ task definition."
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "There is a conflict with the name as there is another task\
            \ definition using it."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Unexpected error, The task definition may not have been created."
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "TaskDefinition"
...
definitions:
  TaskDefinition:
    type: "object"
    properties:
      id:
        type: "string"
        example: "49a9b246-3b5a-4c73-a294-816b25a2ff6c"
        description: "Task Definition id"
        readOnly: true
      name:
        type: "string"
        example: "Security Exam"
        description: "Task Definition name"
        maxLength: 100
      priority:
        type: "string"
        example: "High"
        description: "Indicates the level of importance of the defined task"
        enum:
        - "High"
        - "Medium"
        - "Low"
      type:
        type: "string"
        example: "self-study"
        description: "This field allows classifying the task definition in a specific\
          \ group"
        maxLength: 60
      audience:
        type: "array"
        example:
        - "GENERAL"
        - "JAVA"
        - "DEVELOPMENT"
        description: "This field will contain list of values that allows to identify\
          \ the group of people that are going to execute the task"
        items:
          type: "string"
      approvedBy:
        type: "string"
        example: "buddy"
        description: "This file indicates who is in charge of approving the task when\
          \ it's completed"
      description:
        type: "string"
        example: "Take the Endava security exam available here http://endava.com/security-exam"
        description: "Contains the details about the task"
      files:
        type: "array"
        description: "set of unique documents ids should be included as part of the\
          \ task definition"
        items:
          $ref: "#/definitions/Document"
    description: "Task Definition is a task template which describes an activity that\
      \ needs to be done in a ramp-up process for a particular audience."
    example:
      audience:
      - "GENERAL"
      - "JAVA"
      - "DEVELOPMENT"
      name: "Security Exam"
      approvedBy: "buddy"
      description: "Take the Endava security exam available here http://endava.com/security-exam"
      files:
      - name: "myFile.txt"
        link: "http://documentserver/docs/5aea3961-a858-45ce-95cf-2b0f929dde93"
        doc-id: "5aea3961-a858-45ce-95cf-2b0f929dde93"
      - name: "myFile.txt"
        link: "http://documentserver/docs/5aea3961-a858-45ce-95cf-2b0f929dde93"
        doc-id: "5aea3961-a858-45ce-95cf-2b0f929dde93"
      id: "49a9b246-3b5a-4c73-a294-816b25a2ff6c"
      priority: "High"
      type: "self-study"
...
  Document:
    type: "object"
    properties:
      doc-id:
        type: "string"
        example: "5aea3961-a858-45ce-95cf-2b0f929dde93"
        description: "unique identifier for a document"
      name:
        type: "string"
        example: "myFile.txt"
        description: "file name"
      link:
        type: "string"
        example: "http://documentserver/docs/5aea3961-a858-45ce-95cf-2b0f929dde93"
        description: "URL to download the file"
    example:
      name: "myFile.txt"
      link: "http://documentserver/docs/5aea3961-a858-45ce-95cf-2b0f929dde93"
      doc-id: "5aea3961-a858-45ce-95cf-2b0f929dde93"
  Id:
    type: "object"
    properties:
      id:
        type: "string"
        example: "5aea3961-a858-45ce-95cf-2b0f929dde93"
    example:
      id: "5aea3961-a858-45ce-95cf-2b0f929dde93"
  Error:
    type: "object"
    properties:
      name:
        type: "string"
        example: "HTTP error name"
        description: "Name of the error"
      detail:
        type: "string"
        example: "Error detailed description"
        description: "Detailed description of the error"
...